// src/app/models/visitante.model.ts
export class Visitante {
    "cod_visitantes": number;
    "dni_visitantes": number;
    "nombre_visitante": string;
    "apellido_visitante": string;
    "fecha_visitante": string;
    "corro_visitante": string;
    "telefono_visitantes": number;
    "cod_usuario": string;
    "cod_propitario": string;
}